separated image upload in a new component
